+++ core/src/main/java/org/example/core/repository/UserRepo.java
@@ -9,23 +9,23 @@ import java.util.Map;
 import java.util.Optional;
 
 @Repository
-public class UserRepo implements CRUDRepo<Integer, User> {
+public class UserRepo implements CRUDRepo<Long, User> {
 
-    private final Map<Integer, User> userMap = DB.userMap;
+    private final Map<Long, User> userMap = DB.userMap;
 
-    public Optional<Map.Entry<Integer, User>> save(User user) {
+    public Optional<Map.Entry<Long, User>> save(User user) {
         if (isFieldsIncorrect(user)) {
             return Optional.empty();
         }
 
-        int newId = userMap.size() + 1;
+        long newId = userMap.size() + 1;
         user.setId(newId);
         userMap.put(newId, user);
         return Optional.of(Map.entry(newId, user));
     }
 
     @Override
-    public Optional<Map.Entry<Integer, User>> getById(Integer id) {
+    public Optional<Map.Entry<Long, User>> getById(Long id) {
         if (userMap.get(id) == null) {
             return Optional.empty();
         }
@@ -33,19 +33,19 @@ public class UserRepo implements CRUDRepo<Integer, User> {
     }
 
     @Override
-    public Map<Integer, User> getAll() {
+    public Map<Long, User> getAll() {
         return Collections.unmodifiableMap(userMap);
     }
 
     @Override
-    public Optional<Map.Entry<Integer, User>> update(User user) {
-        Integer id =
+    public Optional<Map.Entry<Long, User>> update(User user) {
+        Long id =
                 userMap.entrySet()
                         .stream()
                         .filter(entry -> entry.getValue().equals(user))
                         .map(Map.Entry::getKey)
                         .findFirst()
-                        .orElse(0);
+                        .orElse(0L);
 
         if (id == 0 || isFieldsIncorrect(user)) {
             return Optional.empty();
@@ -57,7 +57,7 @@ public class UserRepo implements CRUDRepo<Integer, User> {
     }
 
     @Override
-    public void deleteById(Integer id) {
+    public void deleteById(Long id) {
         userMap.remove(id);
     }
