+++ core/src/main/java/org/example/core/service/EventService.java
@@ -8,23 +8,23 @@ import java.util.Map;
 public class EventService {
     private final EventRepo eventRepo = new EventRepo();
 
-    public Map.Entry<Integer, Event> save(Event event) {
-        return eventRepo.save(event).orElseGet(() -> Map.entry(0, new Event()));
+    public Map.Entry<Long, Event> save(Event event) {
+        return eventRepo.save(event).orElseGet(() -> Map.entry(0L, new Event()));
     }
 
-    public Map.Entry<Integer, Event> getById(Integer id) {
-        return eventRepo.getById(id).orElseGet(() -> Map.entry(0, new Event()));
+    public Map.Entry<Long, Event> getById(Long id) {
+        return eventRepo.getById(id).orElseGet(() -> Map.entry(0L, new Event()));
     }
 
-    public Map<Integer, Event> getAll() {
+    public Map<Long, Event> getAll() {
         return eventRepo.getAll();
     }
 
-    public Map.Entry<Integer, Event> update(Event event) {
-        return eventRepo.update(event).orElseGet(() -> Map.entry(0, new Event()));
+    public Map.Entry<Long, Event> update(Event event) {
+        return eventRepo.update(event).orElseGet(() -> Map.entry(0L, new Event()));
     }
 
-    public void deleteById(Integer id) {
+    public void deleteById(Long id) {
         eventRepo.deleteById(id);
     }
 }
