+++ core/src/main/java/org/example/core/service/TicketService.java
@@ -8,23 +8,23 @@ import java.util.Map;
 public class TicketService {
     private final TicketRepo ticketRepo = new TicketRepo();
 
-    public Map.Entry<Integer, Ticket> save(Ticket ticket) {
-        return ticketRepo.save(ticket).orElseGet(() -> Map.entry(0, new Ticket()));
+    public Map.Entry<Long, Ticket> save(Ticket ticket) {
+        return ticketRepo.save(ticket).orElseGet(() -> Map.entry(0L, new Ticket()));
     }
 
-    public Map.Entry<Integer, Ticket> getById(Integer id) {
-        return ticketRepo.getById(id).orElseGet(() -> Map.entry(0, new Ticket()));
+    public Map.Entry<Long, Ticket> getById(Long id) {
+        return ticketRepo.getById(id).orElseGet(() -> Map.entry(0L, new Ticket()));
     }
 
-    public Map<Integer, Ticket> getAll() {
+    public Map<Long, Ticket> getAll() {
         return ticketRepo.getAll();
     }
 
-    public Map.Entry<Integer, Ticket> update(Ticket ticket) {
-        return ticketRepo.update(ticket).orElseGet(() -> Map.entry(0, new Ticket()));
+    public Map.Entry<Long, Ticket> update(Ticket ticket) {
+        return ticketRepo.update(ticket).orElseGet(() -> Map.entry(0L, new Ticket()));
     }
 
-    public void deleteById(Integer id) {
+    public void deleteById(Long id) {
         ticketRepo.deleteById(id);
     }
 }
